requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.6
djangorestframework==3.7.1
flake8==3.4.1
httpie==0.9.9
idna==2.6
isort==4.2.15
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
